Avoid reverse DNS lookups when I2Addr is created/set.
It seems the intention was for addr->node to be set lazily when first
required, performing a reverse DNS lookup in the process. However, the
calling of _I2AddrSetNodePort from I2AddrBySAddr and I2AddrSetSAddr
defeats the point and results in redundant lookups when applications
never want to lookup the node name or port name.

Therefore, remove the calls to _I2AddrSetNodePort from I2AddrBySAddr
and I2AddrSetSAddr. The functions that look at addr->node and
addr->port correctly use addr->node_set and addr->port_set and call
_I2AddrSetNodePort as required.